fix mobile visibility affordances for tables
I think someone thought that "display: none !important;" was getting inherited from some of the original visibility styles, when really they were being overwritten by the table styles. For example: if you put "hide-on-phones" on a table, it would not show up because there was no in the phone media query to override "table.hide-on-phones { display: table !important; }".