Add Error, Small to 100% width media query.
Currently (in 3.1.1) you can add a column class to `small` tags in order to size them to match the corresponding `input` (this is for situations where one is not doing `row` classes, but just using the number classes to size/arrange form fields, eg, `<input class="six">`). However, the `small` tag (with corresponding `error` class) then does not get full width on the Small screen size. I believe this Pull Request should fix that.

However, I would encourage the Zurb folks to double check that I've not missed anything as this particular bit of Foundation scss is less familiar to me, and I was admittedly focused on fixing my particular use case. So there may be additional changes that you all would wish to make and/or there may be implications to the changes I'm proposing that I'm not aware of.

:)