Fix for topbar dropdown arrow position when height of topbar is increased.
The topbar dropdown arrow has a `top` position of 22px. This is
hardcoded assuming the default topbar height of 45px hasn't changed.
This fix, fixes the position of the topbar dropdown arrow so that its
top position is `$topbar-height / 2`. It also fixes maintains
consistent padding as the topbar height is changed.