Fix for Reveal scroll issue on touch devices
If you open a reveal modal on a touch device and scroll using your
thumb over .reveal-modal-bg (not .reveal-modal itself), it ends up
closing the reveal as soon as your thumb leaves the touch screen (on
touchend).

From a UX standpoint this seems buggy as I didn't tap the background. I
only placed my thumb on it to scroll the reveal modal, not to close it.