Hooked up Karma Runner and restructured the testing framework.
grunt will now default to running tests for both jQuery and zepto
on PhantomJS (both tiny and small window sizes), Chrome, and Firefox.

Additionally, there are several other karma tasks in place in the
gruntfile to accomodate Mac or Win development (including Safari or IE)
as well as full support for "grunt watch" which will compile all the
SASS and then run all the tests as things change.