adds `{{rel}}` helper to make relative paths a little easier to maintain.
In package.json, the names of handlebars helper modules were also added to the keyworks array,  this is a convenience of Assemble to get around having to define the helpers in the Gruntfile. If this is too abstracted we can put the helpers directly in the templates/helpers dir and register them in the Gruntfile.