Update foundation.orbit.js
Since data-orbit-link does work on non-<a> elements when it comes to controlling the slideshow, the same should be true for updating the active link status for those elements (alternatively both should fail on non-<a> elements). Either way, it should be consistent. This proposal does the former, since I find it to be useful and can't see a reason why not.