Replace the <hr> margin difference with its width
By default the bottom margin of the <hr> tag is 1px smaller than the top margin. This results through the default $hr-border-width: 1px; setting. If you change $hr-border-width to another value the margin-bottom still remains the same. This fix resolves this issue, so the bottom margin respond to the $hr-border-width value.