Stack on Small now works with Fade
In the JS, I add the stack_on_small class first. This way if the visitor is on mobile, the display:none style gets applied to all the other elements like bullets before they are even created.

In the SCSS, I force the opacity to be 1 and disable the css animations.