Fix: topbar title h1 disappers behind menu item.
This is only a small issue and I am not sure whether it really matters
in a really live context. When pulling the browser window smaller, the
title disappears behind the Menu item. This can be remedied by giving
the title a little more space. There is still sufficient space for the
Menu item.