Orbit: Fix for links not working on some devices
When a user wraps an image in a link inside Orbit, the link should be set to display:block. I have limited this styles to only links that are direct children of the <li>. We don't want to set all links inside Orbit to be display:block (such as links inside a caption).