Allow clearing to load from closest A not parent
Currently clearing loads images from the current item if it's an `a` tag or from the parent of the current item or failing that it falls back to the item it's self. This means that the only way to load an image from the `a` tag is if the `img` tag is the direct descendant. By changing to use `closest('a')` instead of `parent()` then we gain more flexibility in the HTML we can use.