Fixes empty z-index property to be 'auto'.
Fixes a bug that causes a z-index property to have no value. Rather than using 'null' (which results in no value for the property), 'auto' is used instead.

This bug is a problem if you use a build process like Uncss to trim unused CSS. The build will fail as the empty z-index property isn't valid.