Tabs heavy-handed tabindex setting fix
This change is to remove the tabindex attrs instead of setting them to zero on change of tab.

The problem with having a zero tabindex (on a div for example) is that Chrome highlights the area on click and includes it in the tabbable items, so removing the tabindex is a better option as it reverts the tab section to its default tab behaviour.