Fix broken back button in tabs with deep linking
This change fixes a bug in tabs which have deep linking enabled, whereby
the user, in using the browser back button, would only be able to
navigate backwards as far as the default tab hash location. Attempting
to navigate back to the previous page would result in infinite
redirection back to the default tab. For moreinformation see [this
issue](https://github.com/zurb/foundation/issues/6102)

This change has the tab control remember the entry location upon
initialization, and then whenever the hash changes, it compares the
current location.hash value with the entry location and the default hash
value, in order to determine whether to update the location.hash or
simply allow the navigation to occur without modifying the
location.hash.